Made proper CRUD for deck and card

+++ api/api.py
@@ -82,21 +82,79 @@ class UserRegisterAPI(Resource):
 # properly coded till here
 
 
-class UserOwnDeckList(Resource):
+class UserDeckList(Resource):
 
     def get(self, user_id, api_key):
+        """Used to get the list of decks created by the user
+
+        Args:
+            user_id (int): id of the user (used in the database)
+            api_key (string - 16 chars): used for user api auth everytime a request is made
+        """
 
         if not checkUserValid(user_id=user_id, api_key=api_key):
             invalidUserCred()
 
         user = User.query.filter(User.user_id == user_id).first()
 
-        deck_list = [{'deck_id': deck.deck_id, 'deck_name': deck.deckname,
-                      'public': deck.public, 'no_of_cards': deck.no_of_cards()} for deck in user.decks]
+        deck_list = [
+            {'deck_id': deck.deck_id,
+             'deck_name': deck.deckname,
+             'public': deck.public,
+             'no_of_cards': deck.no_of_cards()
+             }
+            for deck in user.decks
+        ]
 
-        return {"no_of_decks": user.no_of_decks(), "decks": deck_list}
+        return {
+            "no_of_decks": user.no_of_decks(),
+            "decks": deck_list
+        }  # TODO check if this is the best way to send the data back
+
+
+class DeckResource(Resource):
+    """
+    Deck CRUD operation (for decks related to the user)
+    """
+
+    def get(self, user_id, api_key, deck_id):
+        """Used to get information about a certain deck such as its creator, name, visibilty to the public, number of cards. (READ)
+
+        Args:
+            user_id (int): id of the user (used in the database)
+            api_key (string - 16 chars): used for user api auth everytime a request is made
+            deck_id (int): id of the specific deck requested
+
+        Returns:
+            [type]: [description]
+        """
+
+        if not checkUserValid(user_id=user_id, api_key=api_key):
+            invalidUserCred()
+
+        deck = Deck.query.filter(
+            (Deck.deck_id == deck_id) & ((Deck.author_id == user_id) | (Deck.public == True))).first()
+
+        if deck is None:
+            # TODO put proper error
+            return {"error": "No such deck exists for this user"}
+
+        creator = deck.author_id == user_id
+
+        return {
+            'creator': creator,
+            'deck_id': deck.deck_id,
+            'deck_name': deck.deckname,
+            'public': deck.public,
+            'no_of_cards': deck.no_of_cards()
+        }  # TODO check if this is the best way to send the data back
 
     def post(self):
+        """Used to enter into of a new deck into the database (CREATE) 
+
+        Returns:
+            [type]: [description]
+        """
         args = deck_creation_parser.parse_args()
 
         user_id = args["user_id"]
@@ -111,36 +169,29 @@ class UserOwnDeckList(Resource):
             (Deck.author_id == user_id) & (Deck.deckname == deckname)).first()
 
         if deck_present:
-            print('Present')
+            # TODO add an error code here for duplicate name for that particular user
+            print('Duplicate deck name present')
         else:
-            print('Not present')
-            new_deck = Deck(author_id=user_id, deckname=deckname,
-                            public=bool(public))
+            new_deck = Deck(author_id=user_id,
+                            deckname=deckname, public=bool(public))
             db.session.add(new_deck)
             db.session.commit()
-            print('Deck commited properly')
 
-        return {'sat': 'hi'}
-
-    def delete(self, user_id, api_key, deck_id):
-
-        if not checkUserValid(user_id=user_id, api_key=api_key):
-            invalidUserCred()
-
-        d = Deck.query.filter(Deck.deck_id == deck_id).first()
-        db.session.delete(d)
-        db.session.commit()
-
-        return {'sta': 'good'}
+        return {'sat': 'hi'}  # TODO send back proper response
 
     def put(self):
+        """Used for updating the information of the exisiting deck (UPDATE)
+
+        Returns:
+            [type]: [description]
+        """
         args = deck_updation_parser.parse_args()
 
         user_id = args["user_id"]
         api_key = args["api_key"]
+        deck_id = args["deck_id"]
         deckname = args["deckname"]
         public = args["public"]
-        deck_id = args["deck_id"]
 
         if not checkUserValid(user_id=user_id, api_key=api_key):
             invalidUserCred()
@@ -148,19 +199,68 @@ class UserOwnDeckList(Resource):
         u_deck = Deck.query.filter(
             (Deck.deck_id == deck_id) & (Deck.author_id == user_id)).first()
 
+        if not u_deck:  # deck does not exist
+            # TODO give proper error message
+            return 'Deck does not exist'
+
         if deckname:
+
+            duplicate_deck_name = Deck.query.filter(
+                (Deck.author_id == user_id) & (Deck.deckname == deckname)).first()
+
+            if duplicate_deck_name:  # if a deck with that name already exists
+                # TODO give proper error message
+                return 'Deck name should not be duplicate'
+
             u_deck.deckname = deckname
+
         u_deck.public = bool(public)
 
         db.session.add(u_deck)
         db.session.commit()
 
-        return 'Success'
+        return 'Success'  # TODO send back proper response
+
+    def delete(self, user_id, api_key, deck_id):
+        """Used to delete the specific deck requested (DELETE)
+
+        Args:
+            user_id (int): id of the user (used in the database)
+            api_key (string - 16 chars): used for user api auth everytime a request is made
+            deck_id (int): id of the specific deck requested
+
+        Returns:
+            [type]: [description]
+        """
+        if not checkUserValid(user_id=user_id, api_key=api_key):
+            invalidUserCred()
+
+        deck_exists = Deck.query.filter(
+            (Deck.author_id == user_id) & (Deck.deck_id == deck_id)).first()
+
+        if not deck_exists:
+            # TODO give proper error message
+            return 'Deck either not there or does not belong to the user'
+
+        db.session.delete(deck_exists)
+        db.session.commit()
+
+        return {'sta': 'good'}  # TODO send back proper response
 
 
-class UserOwnDeckCards(Resource):
+class DeckCardList(Resource):
 
     def get(self, user_id, api_key, deck_id):
+        """Used to get the list of cards of a particular deck
+
+        Args:
+            user_id (int): id of the user (used in the database)
+            api_key (string - 16 chars): used for user api auth everytime a request is made
+            deck_id (int): id of the specific deck requested
+
+        Returns:
+            [type]: [description]
+        """
 
         if not checkUserValid(user_id=user_id, api_key=api_key):
             invalidUserCred()
@@ -169,25 +269,75 @@ class UserOwnDeckCards(Resource):
             (Deck.deck_id == deck_id) & ((Deck.author_id == user_id) | (Deck.public == True))).first()
 
         if deck is None:
+            # TODO give proper error message
             return {"error": "No such deck exists for this user"}
 
-        card_list = [{'card_id': card.card_id, 'front': card.front,
-                      'back': card.back} for card in deck.cards]
-
-        owner = deck.author_id == user_id
+        card_list = [
+            {'card_id': card.card_id,
+             'front': card.front,
+             'back': card.back
+             }
+            for card in deck.cards
+        ]
 
         return {
-            'owner': owner,
+            'creator': deck.author_id == user_id,
             'deck_id': deck.deck_id,
             'deck_name': deck.deckname,
             'public': deck.public,
             'no_of_cards': deck.no_of_cards(),
             "cards": card_list
-        }
+        }  # TODO check if this is the best way to send the data back
 
-    pass
+
+class CardResource(Resource):
+    """
+    Card CRUD operation (for cards related to a specific deck) \n
+    Create, Delete, Update operations can only be done by the 'creator' user \n
+    Read is possible for the 'creator' user and also for other users of the deck is public \n
+    """
+
+    def get(self, user_id, api_key, card_id):
+        """Used to get a particular card created by user or part of a public deck (READ)
+
+        Args:
+            user_id (int): id of the user (used in the database)
+            api_key (string - 16 chars): used for user api auth everytime a request is made
+            card_id (int): id of the specific card requested
+
+        Returns:
+            [type]: [description]
+        """
+
+        if not checkUserValid(user_id=user_id, api_key=api_key):
+            invalidUserCred()
+
+        card = Card.query.filter(Card.card_id == card_id).first()
+
+        if not card:
+            # TODO give proper error message
+            return {"error": "No such card exists"}
+
+        if not card.deck.public:
+            if not (card.deck.author_id == user_id):
+                # TODO give proper error message
+                return {"error": "No such card exists for this user"}
+
+        return{
+            'creator': card.deck.author_id == user_id,
+            'deck_id': card.deck.deck_id,
+            'deck_name': card.deck.deckname,
+            'card_id': card.card_id,
+            'card_front': card.front,
+            'card_back': card.back
+        }  # TODO check if this is the best way to send the data back
 
     def post(self):
+        """Used to create a new card to be part of a deck created by user (CREATE)
+
+        Returns:
+            [type]: [description]
+        """
         args = card_creation_parser.parse_args()
 
         user_id = args["user_id"]
@@ -199,14 +349,33 @@ class UserOwnDeckCards(Resource):
         if not checkUserValid(user_id=user_id, api_key=api_key):
             invalidUserCred()
 
+        deck = Deck.query.filter((Deck.deck_id == deck_id) & (
+            Deck.author_id == user_id)).first()
+
+        if not deck:
+            # TODO give proper error message
+            return {"error": "No such deck exists for this user"}
+
+        card = Card.query.filter((Card.deck_id == deck_id) & (
+            (Card.front == front) | (Card.back == back))).first()
+
+        if card:
+            # TODO give proper error message
+            return {"error": "Duplicate Card for this deck"}
+
         new_card = Card(front=front, back=back, deck_id=deck_id)
 
         db.session.add(new_card)
         db.session.commit()
-        print('New Card added properly')
-        pass
+
+        return {'sat': 'hi'}  # TODO send back proper response
 
     def put(self):
+        """Used to update a particular card of a deck (UPDATE)
+
+        Returns:
+            [type]: [description]
+        """
         args = card_updation_parser.parse_args()
 
         user_id = args["user_id"]
@@ -219,28 +388,74 @@ class UserOwnDeckCards(Resource):
         if not checkUserValid(user_id=user_id, api_key=api_key):
             invalidUserCred()
 
-        u_card = Card.query.filter(
+        card = Card.query.filter(
             (Card.deck_id == deck_id) & (Card.card_id == card_id)).first()
 
-        if front:
-            u_card.front = front
-        if back:
-            u_card.back = back
+        if not card:
+            # TODO give proper error message
+            return {"error": "Card does not exist"}
 
-        db.session.add(u_card)
+        if not (card.deck.author_id == user_id):
+            # TODO give proper error message
+            return {"error": "Card does not exist for this user"}
+
+        if front:
+            c = Card.query.filter((Card.deck_id == deck_id)
+                                  & (Card.front == front)).first()
+
+            if c:
+                # TODO give proper error message
+                return {"error": "Card duplaicate with new front"}
+
+            card.front = front
+
+        if back:
+
+            c = Card.query.filter((Card.deck_id == deck_id)
+                                  & (Card.back == back)).first()
+
+            if c:
+                # TODO give proper error message
+                return {"error": "Card duplaicate with new back"}
+
+            card.back = back
+
+        db.session.add(card)
         db.session.commit()
 
+        return {'sat': 'hi'}  # TODO send back proper response
+
     def delete(self, user_id, api_key, deck_id, card_id):
+        """Used to delete a particular card from a deck (DELETE)
+
+        Args:
+            user_id (int): id of the user (used in the database)
+            api_key (string - 16 chars): used for user api auth everytime a request is made
+            deck_id (int): id of the specific deck requested
+            card_id (int): id of the specific card requested
+
+        Returns:
+            [type]: [description]
+        """
 
         if not checkUserValid(user_id=user_id, api_key=api_key):
             invalidUserCred()
 
         c = Card.query.filter((Card.card_id == card_id) &
                               (Card.deck_id == deck_id)).first()
+
+        if not c:
+            # TODO give proper error message
+            return {"error": "Card does not exist"}
+
+        if not c.deck.author_id == user_id:
+            # TODO give proper error message
+            return {"error": "Card does not exist for this user"}
+
         db.session.delete(c)
         db.session.commit()
 
-        return {'sta_delete_card': 'good'}
+        return {'sta_delete_card': 'good'}  # TODO send back proper response
 
 
 class PublicDecks(Resource):
@@ -254,7 +469,6 @@ class PublicDecks(Resource):
         deck_list = [{'deck_id': deck.deck_id, 'deck_name': deck.deckname,
                       'public': deck.public, 'no_of_cards': deck.no_of_cards(), "author": deck.get_author()} for deck in decks]
 
-        print(deck_list)
         return {"no_of_decks": len(deck_list), "decks": deck_list}
 
 
@@ -263,13 +477,6 @@ class GettingCard(Resource):
 
         if not checkUserValid(user_id=user_id, api_key=api_key):
             invalidUserCred()
-
-        # if card_id == 0:
-        #     sd = SolvingDeck(user_id=user_id, deck_id=deck_id,
-        #                      start_time=datetime.now())
-        #     db.session.add(sd)
-        #     db.session.commit()
-        #     print(sd.solve_id)
 
         card = Card.query.filter((Card.deck_id == deck_id) & (
             Card.card_id > card_id)).first()
@@ -360,6 +567,9 @@ class GetDecksAttempted(Resource):
             'user_id': user_id,
             'decks_attempted': [
                 {
+                    'creator': record.solvedecks_r.author_id == user_id,
+                    'deck_id': record.solvedecks_r.deck_id,
+                    'public': record.solvedecks_r.public,
                     'deckname': record.solvedecks_r.deckname,
                     'date': record.start_time.strftime("%d-%b-%Y"),
                     'author': record.solvedecks_r.author.username,
@@ -367,29 +577,4 @@ class GetDecksAttempted(Resource):
                 }
                 for record in r
             ]
-        }
-
-
-class DeckBriefInfo(Resource):
-    def get(self, user_id, api_key, deck_id):
-
-        if not checkUserValid(user_id=user_id, api_key=api_key):
-            invalidUserCred()
-
-        deck = Deck.query.filter(
-            (Deck.deck_id == deck_id) & ((Deck.author_id == user_id) | (Deck.public == True))).first()
-
-        if deck is None:
-            # put proper thing here
-            return {"error": "No such deck exists for thihs user"}
-
-        # may change the variable name later (to author)
-        owner = deck.author_id == user_id
-
-        return {
-            'owner': owner,
-            'deck_id': deck.deck_id,
-            'deck_name': deck.deckname,
-            'public': deck.public,
-            'no_of_cards': deck.no_of_cards()
         }
